feat: add ShapeFactoryClass

+++ src/main/java/radyuk/edu/shape/entity/Point.java
@@ -1,4 +1,4 @@
 package radyuk.edu.shape.entity;
 
-public record Point(float x, float y) {
+public record Point(double x, double y) {
 }
